validatorTask

+++ controllers/tasks/removeTaskById.js
@@ -5,8 +5,8 @@ const {Task} = require("../../models/taskModel");
 
 module.exports = async (req, res) => {
     const { id } = req.params;
-   
-    const result = await Task.findByIdAndDelete(id);
+   const { _id: owner} =req.user
+    const result = await Task.findByIdAndDelete({ _id: id, owner });
     if(!result) {
         throw AppError(404, "Not found")
     }
